demo7/data/requirements.txt
azure-search-documents
PyPDF2
python-dotenv
beautifulsoup4